[CASSH SERVER] v1.7.2 : Split tools
